// resources/assets/src/app/models/api/response/SettingReponse.ts
export class TimeCongig {
    id: number;
    start_morning: string;
    end_afternoon: string;
    end_morning: string;
    start_afternoon: string;
    in_late_threshold: number;
    time_off_registration_threshold: number;
    hr_and_administration_mail: string;
    bom_mail: string;
    time_off_reset_milestone: string;

    constructor() {

    }
}